Improve reuse of `dropna()` by moving the bulk of it to Frame()
